refactor: reorganize category stats functionality into stats module
