+++ app/controller/categories/__init__.py
@@ -1,4 +1,3 @@
 from app.controller.categories.get_categories_controller import GetCategoriesController
-from app.controller.categories.get_category_stats_controller import GetCategoryStatsController
 
-__all__ = ["GetCategoriesController", "GetCategoryStatsController"]
+__all__ = ["GetCategoriesController"]
