+++ app/controller/stats/__init__.py
@@ -1,3 +1,4 @@
+from app.controller.stats.get_category_stats_controller import GetCategoryStatsController
 from app.controller.stats.get_overview_stats_controller import GetOverviewStatsController
 
-__all__ = ["GetOverviewStatsController"]
+__all__ = ["GetCategoryStatsController", "GetOverviewStatsController"]
