+++ app/schemas/category_stats_schema.py
@@ -1,10 +1,9 @@
 from pydantic import BaseModel
-from typing import Optional
 
 
 class CategoryStats(BaseModel):
     name: str
     book_count: int
-    min_price: Optional[float]
-    max_price: Optional[float]
-    avg_price: Optional[float]
+    min_price: float | None
+    max_price: float | None
+    avg_price: float | None
