+++ app/schemas/stats_schema.py
@@ -1,9 +1,8 @@
 from pydantic import BaseModel
-from typing import Dict, Optional
 
 
 class OverviewStats(BaseModel):
     total_books: int
-    avg_price: Optional[float]
-    rating_distribution: Dict[str, int]
+    avg_price: float | None
+    rating_distribution: dict[str, int]
