+++ app/usecases/category/__init__.py
@@ -1,4 +1,3 @@
 from app.usecases.category.get_categories_use_case import GetCategoriesUseCase
-from app.usecases.category.get_category_stats_use_case import GetCategoryStatsUseCase
 
-__all__ = ["GetCategoriesUseCase", "GetCategoryStatsUseCase"]
+__all__ = ["GetCategoriesUseCase"]
