+++ app/usecases/stats/__init__.py
@@ -1,3 +1,4 @@
+from app.usecases.stats.get_category_stats_use_case import GetCategoryStatsUseCase
 from app.usecases.stats.get_overview_stats_use_case import GetOverviewStatsUseCase
 
-__all__ = ["GetOverviewStatsUseCase"]
+__all__ = ["GetCategoryStatsUseCase", "GetOverviewStatsUseCase"]
